complete project

+++ todolist_app/views.py
@@ -4,17 +4,22 @@ from todolist_app.models import TaskList
 from todolist_app.forms import TaskForm
 from django.contrib import messages
 from django.core.paginator import Paginator
+from django.contrib.auth.decorators import login_required
 
 
+@login_required
 def todolist(request):
     if request.method == "POST":
         form = TaskForm(request.POST or None)
         if form.is_valid():
-            form.save()
+            instance = form.save(commit=False)
+            instance.manage = request.user
+            instance.save()
         messages.success(request, "New Task Added!")
         return redirect('todolist')
     else:
-        all_tasks = TaskList.objects.all()
+        # all_tasks = TaskList.objects.all()
+        all_tasks = TaskList.objects.filter(manage=request.user)
         paginator = Paginator(all_tasks, 5)
         page = request.GET.get('pg')
         all_tasks = paginator.get_page(page)
@@ -22,12 +27,17 @@ def todolist(request):
         return render(request, 'todolist.html', {'all_task': all_tasks})
 
 
+@login_required
 def delete_task(request, task_id):
     task = TaskList.objects.get(pk=task_id)
-    task.delete()
+    if task.manage == request.user:
+        task.delete()
+    else:
+        messages.error(request, "Access Restricted, You are not allowed!")
     return redirect('todolist')
 
 
+@login_required
 def edit_task(request, task_id):
     if request.method == "POST":
         task = TaskList.objects.get(pk=task_id)
@@ -42,13 +52,18 @@ def edit_task(request, task_id):
         return render(request, 'edit.html', {'task_obj': task_obj})
 
 
+@login_required
 def complete_task(request, task_id):
     task = TaskList.objects.get(pk=task_id)
-    task.done = True
-    task.save()
+    if task.manage == request.user:
+        task.done = True
+        task.save()
+    else:
+        messages.error(request, "Access Restricted, You are not allowed!")
     return redirect('todolist')
 
 
+@login_required
 def pending_task(request, task_id):
     task = TaskList.objects.get(pk=task_id)
     task.done = False
